+++ src/mocks/offers.js
@@ -1,0 +1,38 @@
+const offers = {
+  offers: [
+    {
+      answer: [
+        {
+          src: `path`,
+          title: `Beautiful &amp; luxurious apartment at great location`,
+          price: `120`,
+          stars: `4`,
+          name: `Apartment`,
+        },
+        {
+          src: `path`,
+          title: `Wood and stone place`,
+          price: `90`,
+          stars: `5`,
+          name: `Apartment`,
+        },
+        {
+          src: `path`,
+          title: `Canal View Prinsengracht`,
+          price: `80`,
+          stars: `4`,
+          name: `Apartment`,
+        },
+        {
+          src: `path`,
+          title: `Nice, cozy, warm big bed apartment`,
+          price: `50`,
+          stars: `3`,
+          name: `Apartment`,
+        },
+      ]
+    }
+  ]
+};
+
+export default offers;
